super changes about rawmaterial

+++ Isomerization.Domain/Task1/CalculationParameters.cs
@@ -12,7 +12,7 @@ public struct CalculationParameters
 
     public double HeatCap { get; init; }
     public double T0 { get; init; }
-    public double[] C0 { get; init; }
+    public decimal[] C0 { get; init; }
     public double Activity { get; init; }
 
     public int MaterialCount => C0.Length;
